move user-loader to a separate file and remove global middleware
