+++ packages/express/__tests__/express-middleware.ts
@@ -1,11 +1,9 @@
 import accountsExpress from '../src';
-import { userLoader } from '../src/express-middleware';
 import * as express from 'express';
 
 jest.mock('express', () => {
   const mockRouter = {
     post: jest.fn(),
-    use: jest.fn(),
     get: jest.fn(),
   };
   return {
@@ -27,7 +25,6 @@ describe('express middleware', () => {
       } as any,
       { path: 'test' }
     );
-    expect(router.use).toHaveBeenCalledTimes(1); // Assigning user provider
     expect(router.post.mock.calls[0][0]).toBe('test/impersonate');
     expect(router.post.mock.calls[1][0]).toBe('test/user');
     expect(router.post.mock.calls[2][0]).toBe('test/refreshTokens');
@@ -44,7 +41,6 @@ describe('express middleware', () => {
       } as any,
       { path: 'test' }
     );
-    expect(router.use).toHaveBeenCalledTimes(1); // Assigning user provider
     expect(router.post.mock.calls[0][0]).toBe('test/impersonate');
     expect(router.post.mock.calls[1][0]).toBe('test/user');
     expect(router.post.mock.calls[2][0]).toBe('test/refreshTokens');
@@ -70,71 +66,11 @@ describe('express middleware', () => {
       } as any,
       { path: 'test' }
     );
-    expect(router.use).toHaveBeenCalledTimes(1); // Assigning user provider
     expect(router.post.mock.calls[0][0]).toBe('test/impersonate');
     expect(router.post.mock.calls[1][0]).toBe('test/user');
     expect(router.post.mock.calls[2][0]).toBe('test/refreshTokens');
     expect(router.post.mock.calls[3][0]).toBe('test/logout');
     expect(router.post.mock.calls[4][0]).toBe('test/:service/authenticate');
     expect(router.get.mock.calls[0][0]).toBe('test/oauth/:provider/callback');
-  });
-});
-
-const user = { id: '1' };
-const accountsServer = {
-  resumeSession: jest.fn(() => user),
-};
-describe('userLoader', () => {
-  beforeEach(() => {
-    jest.clearAllMocks();
-  });
-
-  it('does noting when request has no accessToken', async () => {
-    const provider = userLoader(accountsServer as any);
-    const req = {};
-    const res = {};
-    const next = jest.fn();
-    await provider(req, res, next);
-
-    expect(accountsServer.resumeSession).not.toHaveBeenCalled();
-    expect(req).toEqual({});
-    expect(res).toEqual({});
-    expect(next).toHaveBeenCalledTimes(1);
-  });
-
-  it('load user to req object when access token is present on the headers', async () => {
-    const provider = userLoader(accountsServer as any);
-    const req = {
-      headers: {
-        'accounts-access-token': 'token',
-      },
-    };
-    const reqCopy = { ...req };
-    const res = {};
-    const next = jest.fn();
-    await provider(req, res, next);
-
-    expect(accountsServer.resumeSession).toHaveBeenCalledWith('token');
-    expect(req).toEqual({ ...reqCopy, user, userId: user.id });
-    expect(res).toEqual({});
-    expect(next).toHaveBeenCalledTimes(1);
-  });
-
-  it('load user to req object when access token is present on the body', async () => {
-    const provider = userLoader(accountsServer as any);
-    const req = {
-      body: {
-        accessToken: 'token',
-      },
-    };
-    const reqCopy = { ...req };
-    const res = {};
-    const next = jest.fn();
-    await provider(req, res, next);
-
-    expect(accountsServer.resumeSession).toHaveBeenCalledWith('token');
-    expect(req).toEqual({ ...reqCopy, user, userId: user.id });
-    expect(res).toEqual({});
-    expect(next).toHaveBeenCalledTimes(1);
   });
 });
